fix(ui/dataLoaders): Move plugins sidebar into only collectors configure step Co-authored-by: Daniel Campbell <metalwhirlwind@gmail.com>
